airflow/requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-google
datetime
pandas
pathlib
pyspark
yfinance